+++ binutils-2.17/bfd/elf-eh-frame.c
@@ -917,7 +917,7 @@ _bfd_elf_discard_section_eh_frame
       {
 	if (!ent->cie)
 	  {
-	    ecie = ecies + (unsigned long) ent->cie_inf;
+	    ecie = ecies + (intptr_t) ent->cie_inf;
 	    ent->cie_inf = ecie->cie.cie_inf;
 	  }
 	ent->new_offset = offset;
